improves the removal of the `wiki-edit` button
This was previously done by setting the `wiki` attribute to false. This causes inconsistency because the post is actually still a wiki. 

Setting it to `wiki` false causes the admin wrench menu to think that it's not a wiki. So the wiki button there always show "make wiki". The button still works to toggle wiki on/off but causes confusion since the label never changes. 

This commit changes the way the wiki edit button is removed.