Open shared-edits composer if post is a shared edit
The "edit-post" button in the history modal currently ignores the fact that a post is set to "shared-edit" 

This causes it to open the default edit composer and creates sync issues between users using the default composer and the ones using the shared-edits composer. 

This commit makes it so that the button opens the shared-edit composer if the post is a shared-edit. Otherwise, fallback to the default behavior. 

By default, a wiki post will have a different label for the button. "Edit wiki" If the post is both a wiki and a shared edit, this commit changes the label back to "edit post" instead of "edit wiki" 

The most common use case for wiki + shared-edit is to allow more users to edit it with shared edits, not that they really want it to be a wiki. So, the label change in core only adds confusion. Let's reset it.