DEV: improve interaction between wiki and shared-edit posts (#31)
* add css class to share-edit posts

* improves the removal of the default edit button

Shared-edit adds it's own button. So the default one is not needed.  

There's an API method we can use to remove the button when the post is in shared-edit mode.

* improves the removal of the `wiki-edit` button

This was previously done by setting the `wiki` attribute to false. This causes inconsistency because the post is actually still a wiki. 

Setting it to `wiki` false causes the admin wrench menu to think that it's not a wiki. So the wiki button there always show "make wiki". The button still works to toggle wiki on/off but causes confusion since the label never changes. 

This commit changes the way the wiki edit button is removed.

* shrink the reply button for shared-edit posts

If a post is in shared edit mode and the user can edit it and sees the "shared-edit" button. Show the reply button without a label. Otherwise, we have two buttons with labels.

* Open shared-edits composer if post is a shared edit

The "edit-post" button in the history modal currently ignores the fact that a post is set to "shared-edit" 

This causes it to open the default edit composer and creates sync issues between users using the default composer and the ones using the shared-edits composer. 

This commit makes it so that the button opens the shared-edit composer if the post is a shared-edit. Otherwise, fallback to the default behavior. 

By default, a wiki post will have a different label for the button. "Edit wiki" If the post is both a wiki and a shared edit, this commit changes the label back to "edit post" instead of "edit wiki" 

The most common use case for wiki + shared-edit is to allow more users to edit it with shared edits, not that they really want it to be a wiki. So, the label change in core only adds confusion. Let's reset it.